Feature: Generic data model and Supabase data service
A generic data template model has been implemented. This model can be
extended to other classes, such as User. It includes the TABLE_NAME
and DATABASE_MAPPING properties for a more efficient database access.

The SupabaseService has been implemented to support general data
fetching. It currently supports the fetching of a single object and
an array of objects. There is no need to implement the data fetching
logic repeatedly for each model.

- Added SupabaseService
- Added GenericModel
- Added User model
- Added data fetching tests during Homepage initialization